requirements.txt
twscrape
pandas
nltk
python-dotenv
